IDVA6-532 Code refactored as suggested by SonarQube
