+++ src/routerDispatch.ts
@@ -1,6 +1,6 @@
 // Do Router dispatch here, i.e. map incoming routes to appropriate router
 import { Application, Request, Response } from "express";
-import yourCompaniesRouter from "./routers/yourCompaniesRouter";
+import yourCompaniesRouter from "./routers/router";
 import { ERROR_400_TEMPLATE, LANDING_URL } from "./constants";
 
 const routerDispatch = (app: Application) => {
