+++ src/server.ts
@@ -5,8 +5,8 @@ import logger from "./lib/Logger";
 import app from "./app";
 
 // start the HTTP server
-const httpServer = http.createServer(app);
-httpServer.listen(process.env.NODE_PORT, () => {
+const server = http.createServer(app);
+server.listen(process.env.NODE_PORT, () => {
     console.log(`Server started at: ${process.env.NODE_HOSTNAME}:${process.env.NODE_PORT}`);
 }).on("error", err => {
     logger.error(`${err.name} - HTTP Server error: ${err.message} - ${err.stack}`);
@@ -27,4 +27,4 @@ if (process.env.NODE_SSL_ENABLED === "ON") {
     });
 }
 
-export default httpServer;
+export default server;
